migrate TimeVariantsSpec into GenSpec
